pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "md_to_cooklang"
version = "0.1.0"
description = "Convert markdown files to cooklang format using OpenAI"
requires-python = ">=3.8"
dependencies = [
    "openai>=1.0.0",
    "click>=8.0.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0"
]

[project.scripts]
md-to-cooklang = "md_to_cooklang.cli:main"

[tool.hatch.build]
include = [
    "src/md_to_cooklang/*.py",
    "src/md_to_cooklang/*.md"
]

[tool.hatch.build.targets.wheel]
packages = ["src/md_to_cooklang"]
